Change MtTileset scratchpad to a grid<uint8_t>
As the MetaTile Tileset is now 256 MetaTiles in size, the scratchpad/map
can now be stored in a byte grid.

This commit is in preparation of the MetaTile map editor, which will
also use a grid<uint8_t> to store the map data.

The MtGraphicsScene selection grid is still a grid<uint16_t> as we need
a way to differentiate between solid and transparent tiles.

+++ src/gui-qt/metatiles/mttileset/resourcelist.cpp
@@ -72,7 +72,7 @@ void ResourceList::do_addResource(int settingIndex, const std::string& filename)
 
         MT::MetaTileTilesetInput tilesetInput;
         tilesetInput.name = name.toStdString();
-        tilesetInput.scratchpad = grid<uint16_t>(16, 16, ResourceItem::DEFAULT_SCRATCHPAD_TILE);
+        tilesetInput.scratchpad = grid<uint8_t>(16, 16, ResourceItem::DEFAULT_SCRATCHPAD_TILE);
         MT::saveMetaTileTilesetInput(tilesetInput, filename);
     }
 }
